Remove DWORD casts in Hook func

+++ src/main.cpp
@@ -10,10 +10,10 @@ void Init()
     SetTimeSinceLastUpdate();
 
     // Hook various calls
-    Hook(POST_INIT_DEALLOC_CALL_ADDR, (DWORD) PostInitDealloc_Hook, 5);
-    Hook(CHECK_FOR_SYNC_CALL_ADDR, (DWORD) CheckForSync_Hook, 5);
-    Hook(OBJ_UPDATE_CALL_ADDR, (DWORD) SPObjUpdate_Hook, 6);
-    Hook(WAYPOINT_CHECK_CALL_ADDR, (DWORD) WaypointCheck_Hook, 5);
+    Hook(POST_INIT_DEALLOC_CALL_ADDR, PostInitDealloc_Hook, 5);
+    Hook(CHECK_FOR_SYNC_CALL_ADDR, CheckForSync_Hook, 5);
+    Hook(OBJ_UPDATE_CALL_ADDR, SPObjUpdate_Hook, 6);
+    Hook(WAYPOINT_CHECK_CALL_ADDR, WaypointCheck_Hook, 5);
 
     static PVOID projectilesPerFireHookPtr = GetProjectilesPerFire_Hook;
     SetPointer(PROJECTILES_PER_FIRE_CALL_ADDR, &projectilesPerFireHookPtr);
